+++ CHANGELOG.md
@@ -1,5 +1,8 @@
 ## Latest changes
 * Extended documentation
+* Increase testcoverag
+* Fixed calculateLateralDimensions()
+* Extended extractSituation() by timeIndex parameter
 
 ## Release 1.0.0
 * Initial release