+++ CHANGELOG.md
@@ -1,4 +1,5 @@
 ## Latest changes
+* Added support for Clang 5 and Clang 6
 * Added checks to enforce increasing time indices
 * Consider timeIndex=0 as error